feat: ⚰️ (features) ドラック可能な`Member`の作成
